pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentforge"
version = "0.1.0"
description = "AI 에이전트 시대의 선택장애 해결 프레임워크 — What to Build in the Agent Era"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AgentForge Contributors" },
]
keywords = ["ai", "agents", "claude", "llm", "developer-tools", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
]

[project.scripts]
agentforge = "agentforge.cli:main"

[project.urls]
Homepage = "https://github.com/agentforge/agentforge"
Repository = "https://github.com/agentforge/agentforge"
Issues = "https://github.com/agentforge/agentforge/issues"

[tool.hatch.build.targets.wheel]
packages = ["agentforge"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=agentforge --cov-report=term-missing"
tmp_path_retention_policy = "failed"
